\docType{package}
\name{canceR-package}
\alias{canceR-package}
\title{Models of non-myeloid cancerss}
\description{
  This package contains functions that compute the right
  hand sides of ODEs of models relevant to nonmyeloid
  cancers.
}
\details{
  \tabular{ll}{ Package: \tab \pkg{CanceR}\cr Type: \tab
  Package\cr Depends: \tab deSolve\cr Suggests: \tab bbmle
  \cr License: \tab GPL-2\cr LazyLoad: \tab yes\cr
  LazyData: \tab yes\cr URL: \tab
  \url{http://epbi-radivot.cwru.edu/canceR.html}\cr }
}
\note{
  This work was supported by the National Cancer Institute
  and Tufts Integrative Cancer Biology Program under
  U54CA149233-029689.
}
\author{
  Tom Radivoyevitch (\email{txr24@case.edu})
}
\seealso{
  \code{\link{hahnfeldt99}}
}
\keyword{package}

