\name{canceR}
\alias{
canceR
}

\title{
Main function of canceR package.
}
\description{
This function get Studies from cbio portal (\url{http://www.cbioportal.org/}) and list them in starting window. User can select easily multiple studies to get specific data.
}



\references{
canceR: A Graphical User Interface for accessing and modeling the MSKCC Cancer Genomics Data.
}
\author{
Karim Mezhoud Nuclear Safety Department, Nuclear Science Center of Tunisia.
}


\examples{
\dontrun{
canceR()

}
}