% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canceR_pdf.R, R/canceR_vignette.R
\name{canceR_Vignette}
\alias{canceR_Vignette}
\title{open pdf vignette}
\usage{
canceR_Vignette()

canceR_Vignette()
}
\value{
open pdf vignette

open pdf vignette
}
\description{
open pdf vignette

open pdf vignette
}
\examples{
\dontrun{
canceR_Vignette()
}

\dontrun{
canceR_Vignette()
}

}
