% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogMut.R
\name{dialogMut}
\alias{dialogMut}
\title{Dialog bos to set returned Mutation information}
\usage{
dialogMut(title, question, entryInit, entryWidth = 40,returnValOnCancel = "ID_CANCEL")
}
\arguments{
\item{title}{title of the table}

\item{question}{question}

\item{entryInit}{entryInit}

\item{entryWidth}{40}

\item{returnValOnCancel}{"ID_CANCEL"}
}
\value{
a check box with mutations variables
}
\description{
Dialog bos to set returned Mutation information
}
\examples{
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
\dontrun{
dialogMut("title", "question", "entryInit", entryWidth = 40, returnValOnCancel = "ID_CANCEL")
}
}

