% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogOptionCircos.R
\name{dialogOptionCircos}
\alias{dialogOptionCircos}
\title{Checkbox to select dimensions}
\usage{
dialogOptionCircos()
}
\value{
a checkbox with all dimensions
}
\description{
Checkbox to select dimensions
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/Circos.rds", sep=""))
\dontrun{
dialogOptionCircos()
#getCircos(dimension ="All")
}
}
