% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogOptionGSEAlm.R
\name{dialogOptionGSEAlm}
\alias{dialogOptionGSEAlm}
\title{Dialogbox to select variables from Clinical data}
\usage{
dialogOptionGSEAlm(k,ClinicalData)
}
\arguments{
\item{k}{integer 1}

\item{ClinicalData}{dataframe with clinical variables}
}
\value{
permutaion value, p-value, coVariables
}
\description{
Dialogbox to select variables from Clinical data
}
\examples{
#data(ClinicalData)
\dontrun{
getOptionGSEAlm()
}
}
