% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogPlotOption_SkinCor.R
\name{dialogPlotOption_SkinCor}
\alias{dialogPlotOption_SkinCor}
\title{Checkbox to select variables for plotting}
\usage{
dialogPlotOption_SkinCor(s)
}
\arguments{
\item{s}{integer number of Studies}
}
\value{
Dialog box with setting of correlation method
}
\description{
Checkbox to select variables for plotting
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/gbm_tcgaPlotTwoGenProf.rds", sep=""))
\dontrun{
dialogPlotOption_SkinCor(1)
}
}
