% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogSelectFiles_GSEA.R
\name{dialogSelectFiles_GSEA}
\alias{dialogSelectFiles_GSEA}
\title{Dialog Box to  Select GCT, CLS, GMT and output Files for GSEA-R (Broad Institute)}
\usage{
dialogSelectFiles_GSEA()
}
\value{
A vector with files paths
}
\description{
Dialog Box to  Select GCT, CLS, GMT and output Files for GSEA-R (Broad Institute)
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
dialogSelectFiles_GSEA()
}


}
