% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogSummary_GSEA.R
\name{dialogSummary_GSEA}
\alias{dialogSummary_GSEA}
\title{Dialog Box to specify phenotype (variable) used in last GSEA-R to get Summary Results. This function ask the user to specify the phenotype (variable).}
\usage{
dialogSummary_GSEA(Variable,returnValOnCancel ="ID_CANCEL")
}
\arguments{
\item{Variable}{phenotype}

\item{returnValOnCancel}{"ID_CANCEL"}
}
\value{
variables
}
\description{
Dialog Box to specify phenotype (variable) used in last GSEA-R to get Summary Results. This function ask the user to specify the phenotype (variable).
}
\examples{
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
\dontrun{
#Run.GSEA()
#getSummaryGSEA()
}
}

