% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayInTable.R
\name{displayInTable}
\alias{displayInTable}
\title{Display matrix in tcltk table}
\usage{
displayInTable(tclarray,title="",height=-1,width=-1,nrow=-1,ncol=-1)
}
\arguments{
\item{tclarray}{a dataframe}

\item{title}{title of the table}

\item{height}{-1}

\item{width}{-1}

\item{nrow}{-1}

\item{ncol}{-1}
}
\value{
display a Table
}
\description{
Display matrix in tcltk table
}
\examples{
data(ClinicalData)
\dontrun{
getInTable(Table= ClinicalData, title= "Clinical Data")
}
}
