\name{getClinicalDataMatrix}
\alias{getClinicalDataMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getClinicalDataMatrix
}
\description{
Load clinical data from file.
}
\usage{
getClinicalDataMatrix()
}
\value{
Value
}
\examples{

## Not run
##getClinicalDataMatrix()
## End(Not run)
}