% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geteSet.R
\name{geteSet}
\alias{geteSet}
\title{Built Expression Set (eSet) from profile data.}
\usage{
geteSet()
}
\value{
ExpressionSet
}
\description{
Built Expression Set (eSet) from profile data.
}
\examples{
 f <- 9
 \dontrun{
 readRDS(paste(path.package("canceR"),"/extdata/rdata/prad_michPhenoTest1021.rds", sep=""))
 geteSet()
 }

}
