% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCTCLSExample.R
\name{getGCTCLSExample}
\alias{getGCTCLSExample}
\title{get GCT and CLS example files.}
\usage{
getGCTCLSExample()
}
\value{
GCT and CLS files
}
\description{
get GCT and CLS example files.
}
\examples{
## Load workspace
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
getGCTCLSExample()
}
}
