% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneExpMatrix.R
\name{getGeneExpMatrix}
\alias{getGeneExpMatrix}
\title{get matrix with gene expression from file}
\usage{
getGeneExpMatrix()
}
\value{
dataframe of gene expression
}
\description{
get matrix with gene expression from file
}
\examples{
myGlobalEnv <- new.env(parent = emptyenv())
\dontrun{
readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
getGeneExpMatrix()
}
}
