% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneList.R
\name{getGeneList}
\alias{getGeneList}
\title{User needs to specify which gene is interesting to get genomic cancer data. The gene must be with Symbol and one gene by line.}
\usage{
getGeneList()
}
\value{
Gene list path of file
}
\description{
User needs to specify which gene is interesting to get genomic cancer data. The gene must be with Symbol and one gene by line.
}
\examples{
myGlobalEnv <- new.env(parent = emptyenv())
\dontrun{
getGeneList()
 }
}

