% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneListExample.R
\name{getGeneListExample}
\alias{getGeneListExample}
\title{get Gene List from examples. User can select one from available gene list}
\usage{
getGeneListExample()
}
\value{
Gene list path of file
}
\description{
get Gene List from examples. User can select one from available gene list
}
\examples{
myGlobalEnv <- new.env(parent = emptyenv())
\dontrun{
getGeneListExample() 
}
}

