\name{getGenesClassifier}
\alias{getGenesClassifier}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Ranking Genes by Diseases.
}
\description{
refers to geNetClassifier package.
}

\seealso{
geNetClassifier package
}


\examples{
## Not run
##library(canceR)
##library(Biobase)
##library(geNetClassifier)
## Load workspace
#load(paste(path.package("canceR"),"/data/brca_tcgaClassifier223.RData", sep=""))
##getGenesClassifier()
## End(Not run)
}