\name{getGenesTree_MultipleCases}
\alias{getGenesTree_MultipleCases}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getGenesTree_MultipleCases
}
\description{
Get  successively trees of genes list for multiple cases
}
\value{
Value
}
\examples{
## Not run
##library(canceR)
##load(paste(path.package("canceR"),"/data/brca_tcgaClassifier223.RData", sep=""))
##getGenesTree_MultipleCases()
## End(Not run)
}

