\name{getGenesTree_SingleCase}
\alias{getGenesTree_SingleCase}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getGenesTree_SingleCase
}
\description{
classify genes in tree for two phenotypes in the same case(disease).
}
\usage{
getGenesTree_SingleCase()
}

\value{
Value
}
\examples{
## Not run
##library(canceR)
##load(paste(path.package("canceR"),"/data/brca_tcgaClassifier223.RData", sep=""))
##getGenesTree_SingleCase()
##end(Not run)
}