% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenProfs.R
\name{getGenProfs}
\alias{getGenProfs}
\title{Get Genetic Profile from selected Studies}
\usage{
getGenProfs()
}
\value{
dataframe with genetic profil
}
\description{
Get Genetic Profile from selected Studies
}
\examples{

cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
