% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSEAlm_Diseases.R
\name{getGSEAlm_Diseases}
\alias{getGSEAlm_Diseases}
\title{get GSEA linear modeling by studies (diseases)}
\usage{
getGSEAlm_Diseases()
}
\value{
a dataframe with annotation (GO, BP)
}
\description{
get GSEA linear modeling by studies (diseases)
}
\examples{
 load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
 \dontrun{
 load(paste(.libPaths(),"/canceR/data/ucec_tcga_pubGSEA1021.RData", sep=""))
 getGSEAlm_Diseases
}
}

