% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInTable.R
\name{getInTable}
\alias{getInTable}
\title{get dataframe in TK/TCL table}
\usage{
getInTable(table,title)
}
\arguments{
\item{table}{Dataframe}

\item{title}{string a title of the table}
}
\value{
display a Table
}
\description{
get dataframe in TK/TCL table
}
\examples{
#data(ClinicalData)
\dontrun{
getInTable(Table= ClinicalData, title= "Clinical Data")
}
}
