% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMegaProfData.R
\name{getMegaProfData}
\alias{getMegaProfData}
\title{Get profile data for more than 500 genes list.}
\usage{
getMegaProfData(MegaGeneList,k)
}
\arguments{
\item{MegaGeneList}{Genelist >500}

\item{k}{integer number of studies}
}
\value{
dataframewith profile data
}
\description{
Get profile data for more than 500 genes list.
}
\examples{
myGlobalEnv <- new.env(parent = emptyenv())
readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcgaGSEAlm1021.rds", sep=""))
\dontrun{
getMegaProfData(myGlobaEnv$MegaGeneList,1) 
}

}
