% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigDB.R
\name{getMSigDB}
\alias{getMSigDB}
\title{Reduce MSigDB size for only gene list}
\usage{
getMSigDB(eSet, k)
}
\arguments{
\item{eSet}{Expression Set}

\item{k}{integer Number of studies}
}
\value{
MSigDB for user gene List
}
\description{
Reduce MSigDB size for only gene list
}
\examples{
 d <- 7
\dontrun{
setWorkspace()
getMSigDB(eSet = myGlobalEnv$eSetClassifier,k = 1)
}

}
