% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigDBExample.R
\name{getMSigDBExample}
\alias{getMSigDBExample}
\title{get example of .gmt file from MSigDB (Broad Institute)}
\usage{
getMSigDBExample()
}
\value{
path of GMT file
}
\description{
get example of .gmt file from MSigDB (Broad Institute)
}
\examples{
 readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
 \dontrun{
 getMSigDBExample()
 }
}
