% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigDBfile.R
\name{getMSigDBfile}
\alias{getMSigDBfile}
\title{Dialog Box to  Select MSigDB Files from drive}
\usage{
getMSigDBfile()
}
\value{
A path of MSigDB file
}
\description{
Dialog Box to  Select MSigDB Files from drive
}
\examples{
f <- 5+2
\dontrun{
load(paste(path.package("canceR"),"/data/prad_michPhenoTest1021", sep=""))
geteSet()
getMSigDBfile()
}
}

