% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecificMut.R
\name{getSpecificMut}
\alias{getSpecificMut}
\title{get specific Mutation data for multiple genes}
\usage{
getSpecificMut()
}
\value{
a a dataframe with specific mutation informations
}
\description{
get specific Mutation data for multiple genes
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata//ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
getSpecificMut()
}

}
