% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTextWin.R
\name{getTextWin}
\alias{getTextWin}
\title{get text in tcltk windows}
\usage{
getTextWin(text)
}
\arguments{
\item{text}{string}
}
\value{
tcltk windows with text
}
\description{
get text in tcltk windows
}
\examples{
text <- "mytext"
\dontrun{
getTextWin(text) 
}
}

