\name{GSEA.HeatMapPlot2}
\alias{GSEA.HeatMapPlot2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GSEA.HeatMapPlot2
}

\description{
GSEA.HeatMapPlot2
}
\value{
GSEA.HeatMapPlot2
}
\examples{
\dontrun{
library(canceR)
## Load workspace
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
##Run.GSEA()
}
}