% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1Gene_2GenProfs.R
\name{plot_1Gene_2GenProfs}
\alias{plot_1Gene_2GenProfs}
\title{Plotting two genetic profiles for  one Gene}
\usage{
plot_1Gene_2GenProfs()
}
\value{
plot
}
\description{
Plotting two genetic profiles for  one Gene
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/gbm_tcgaPlotTwoGenProf.rds", sep=""))
\dontrun{
plot_1Gene_2GenProfs()
}
}
