\name{plot_2Genes_1GenProf}
\alias{plot_2Genes_1GenProf}

\title{
plot correlation of two genes expressions.
}
\description{
Plotting the correlation between two genes.
}

\examples{
\dontrun{
##library(canceR)
## Load workspace
##load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
##plot_2Genes_1GenProf()
}
}