% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel.R
\name{plotModel}
\alias{plotModel}
\title{model plotting with tcltk}
\usage{
plotModel(plotCommand, title= "TITLE",hscale=1, vscale=1 )
}
\arguments{
\item{plotCommand}{plotcommand}

\item{title}{title of plot}

\item{hscale}{horizintal scale}

\item{vscale}{vertical scale}
}
\value{
plot
}
\description{
model plotting with tcltk
}
\examples{
load(paste(path.package("canceR"),"/data/gbm_tcgaPlotTwoGenProf.RData", sep=""))
\dontrun{
plot_1Gene_2GenProfs()
}
}
