% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.na.R
\name{rbind.na}
\alias{rbind.na}
\title{bind non equal row}
\usage{
rbind.na(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices.}

\item{deparse.level}{integer controlling the construction of labels in 
the case of non-matrix-like arguments (for the default method):
deparse.level = 0 constructs no labels; the default,
deparse.level = 1 or 2 constructs labels from the argument names.}
}
\value{
a data frame with merged rows
}
\description{
bind non equal row
}
\examples{
\dontrun{
row1 <- c("a","b","c","d")
row2 <- c("A", "B", "C")
row3 <- rbind.na(row1, row2)
}


}
