% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWorkspace.R
\name{setWorkspace}
\alias{setWorkspace}
\title{Setting work Directory and output folders.At starting window, user needs to set work directory for output data. The function is foud in File menu.}
\usage{
setWorkspace()
}
\value{
paths of output files
}
\description{
Setting work Directory and output folders.At starting window, user needs to set work directory for output data. The function is foud in File menu.
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
\dontrun{
setWorkspace()
}

}
