% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canceR.R
\name{canceR}
\alias{canceR}
\title{main function}
\usage{
canceR()
}
\value{
open the starting windows with cancer studies
}
\description{
main function
}
\examples{
myGlobalEnv <- new.env(parent = emptyenv())
\dontrun{
canceR()
}
}
