\name{canceR_Vignette}
\alias{canceR_Vignette}

\title{
Get Vignette 
}
\description{
This function open the canceR Vignette pdf file from Help menu.
}
\usage{
canceR_Vignette()
}


\references{
canceR: A Graphical User Interface for accessing and modeling the MSKCC Cancer Genomics Data.
}
\author{
Karim Mezhoud Nuclear Safety Department, Nuclear Science Center. Tunisia
}

\examples{
\dontrun{
canceR_Vignette()

}
}