% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.na.R
\name{cbind.na}
\alias{cbind.na}
\title{bind non equal colunm}
\usage{
cbind.na(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices.}

\item{deparse.level}{integer controlling the construction of labels in 
the case of non-matrix-like arguments (for the default method):
deparse.level = 0 constructs no labels; the default,
deparse.level = 1 or 2 constructs labels from the argument names.}
}
\value{
a data frame with merged columns
}
\description{
bind non equal colunm
}
\examples{
\dontrun{
col1 <- c("a","b","c","d")
col2 <- c("A", "B", "C")
col3 <- cbind.na(col1, col2)
}

}
