% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.na.R
\name{cbind.na}
\alias{cbind.na}
\title{bind non equal colunm}
\usage{
cbind.na(..., deparse.level = 1)
}
\arguments{
\item{...}{...}

\item{deparse.level}{1}
}
\value{
a data frame with merged columns
}
\description{
bind non equal colunm
}
\examples{
col1 <- c("a","b","c","d")
col2 <- c("A", "B", "C")
col3 <- cbind.na(col1, col2)



}

