% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogGeneClassifier.R
\name{dialogGeneClassifier}
\alias{dialogGeneClassifier}
\title{Dialogue Box for gene classifier setting: sample size and postprob threshold}
\usage{
dialogGeneClassifier(Lchecked_Cases,entryWidth = 10,returnValOnCancel = "ID_CANCEL")
}
\arguments{
\item{Lchecked_Cases}{integer with a number of checked cases}

\item{entryWidth}{integer default 10}

\item{returnValOnCancel}{"ID_CANCEL"}
}
\value{
a dataframe with genes classes
}
\description{
Dialogue Box for gene classifier setting: sample size and postprob threshold
}
\examples{
load(paste(path.package("canceR"),"/data/gbm_tcgaPlotTwoGenProf.RData", sep=""))
\dontrun{
getGenesClassifier()
dialogGeneClassifier(1,10,returnValOnCancel = "ID_CANCEL")
}


}

