% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialoggetGeneListMSigDB.R
\name{dialoggetGeneListMSigDB}
\alias{dialoggetGeneListMSigDB}
\title{Multi-select choice of gene sets from loaded MSigDB}
\usage{
dialoggetGeneListMSigDB(MSigDB)
}
\arguments{
\item{MSigDB}{object with MSigDB. A list of genesets}
}
\value{
a dataframe with genes classes
}
\description{
Multi-select choice of gene sets from loaded MSigDB
}
\examples{
z <- 7
\dontrun{
MSigDB <- readLines(paste(.libPaths(),"/canceR/extdata/MSigDB/c5.bp.v4.0.symbols.gmt", sep=""))
dialoggetGeneListMSigDB(MSigDB)
}



}
