% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogMetOption.R
\name{dialogMetOption}
\alias{dialogMetOption}
\title{Dialog Box to set methylation options}
\usage{
dialogMetOption(ProfData, k)
}
\arguments{
\item{ProfData}{adataframe with methylation data}

\item{k}{threshold of silencing gene 0:1}
}
\value{
a dialog box to set methylation option (threshold of silencing gene)
}
\description{
Dialog Box to set methylation options
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
getMetDataMultipleGenes()
#dialogMetOption(ProfData,0.7)
}


}
