% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogOptionPhenoTest.R
\name{dialogOptionPhenoTest}
\alias{dialogOptionPhenoTest}
\title{Checkbox to select variables from clinical data}
\usage{
dialogOptionPhenoTest(eSet)
}
\arguments{
\item{eSet}{Expression Set}
}
\value{
vectors: variables to test Survival status, AGE, p-value
}
\description{
Checkbox to select variables from clinical data
}
\examples{
load(paste(path.package("canceR"),"/data/prad_michPhenoTest1021.RData", sep=""))
\dontrun{
dialogOptionPhenoTest(myGlobalEnv$eSet)
}

}
