\name{dialogPlotOption_SkinCor}
\alias{dialogPlotOption_SkinCor}

\title{
dialogPlotOption_SkinCor
}
\description{
Specifying Skin and Correlation Options.
}



\seealso{
cgdsr package
}

\examples{
\dontrun{
library(canceR)
##Load workspace
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
plot_2Genes_1GenProf()

##dialogPlotOption_SkinCor(2)
}
}