% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogSamplingGSEA.R
\name{dialogSamplingGSEA}
\alias{dialogSamplingGSEA}
\title{Dialog  Box for Sampling patients from expression profile data used for GSEA-R (Broad Institute)}
\usage{
dialogSamplingGSEA(n_checked_GenProf, entryWidth = 10,returnValOnCancel = "ID_CANCEL")
}
\arguments{
\item{n_checked_GenProf}{Number of checked genetic profiles}

\item{entryWidth}{10}

\item{returnValOnCancel}{"ID_CANCEL"}
}
\value{
A vector with sampling size
}
\description{
Dialog  Box for Sampling patients from expression profile data used for GSEA-R (Broad Institute)
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
Run.GSEA()
#dialogSamplingGSEA(1,entryWidth=10,returnValOnCancel = "ID_CANCEL")
}
 
}
