\name{dialogSelectFiles_GSEA}
\alias{dialogSelectFiles_GSEA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Dialog Box to  Select GCT, CLS, GMT and output Files for GSEA-R (Broad Institute) 
}
\description{
This function select GCT, CLS, GMT and output Files for GSEA-R.
}

\examples{
## Not run
library(canceR)
## Load workspace
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
##dialogSelectFiles_GSEA()
## End(Not run)
}