% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogSpecificMut.R
\name{dialogSpecificMut}
\alias{dialogSpecificMut}
\title{dialog box to Specify Mutation using Regular Expression. Search specific mutation using regular expression.}
\usage{
getSpecificMut()
}
\value{
a a dataframe with specific mutation informations
}
\description{
dialog box to Specify Mutation using Regular Expression. Search specific mutation using regular expression.
}
\examples{
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
\dontrun{
getSpecificMut()
}
}
