\name{GeneExpMatrix}
\alias{GeneExpMatrix}
\docType{data}
\title{
GeneExpMatrix
}
\description{
Example of GeneExpMatrix
}
\usage{data("GeneExpMatrix")}
\format{
  A data frame with 958 observations on the following 18 variables.
  \describe{
    \item{\code{BEGAIN}}{a numeric vector}
    \item{\code{CD83}}{a numeric vector}
    \item{\code{CD93}}{a numeric vector}
    \item{\code{CEP164}}{a numeric vector}
    \item{\code{FOXN2}}{a numeric vector}
    \item{\code{IGFBP2}}{a numeric vector}
    \item{\code{IL18}}{a numeric vector}
    \item{\code{KDELR1}}{a numeric vector}
    \item{\code{NCSTN}}{a numeric vector}
    \item{\code{NOTCH2}}{a numeric vector}
    \item{\code{NPY}}{a numeric vector}
    \item{\code{NT5E}}{a numeric vector}
    \item{\code{PARP4}}{a numeric vector}
    \item{\code{SIGLEC1}}{a numeric vector}
    \item{\code{SLC16A2}}{a numeric vector}
    \item{\code{SLC35B1}}{a numeric vector}
    \item{\code{SLC9A2}}{a numeric vector}
    \item{\code{VPS16}}{a numeric vector}
  }
}
\details{
example of gene expression
}
\source{
cbioportal
}
\references{
cbioportal
}
\value{
a dataframe of gene expression
}
\examples{
data(GeneExpMatrix)
## maybe str(GeneExpMatrix) ; plot(GeneExpMatrix) ...
}
\keyword{datasets}
