% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCases.R
\name{getCases}
\alias{getCases}
\title{Get cases for selected Studies. The Cases are the descrption of the samples from patients. The samples can be subdivided by the type of assays as, sequencing, CNA, Mutation, Methylation.}
\usage{
getCases()
}
\value{
a dataframe with cases
}
\description{
Get cases for selected Studies. The Cases are the descrption of the samples from patients. The samples can be subdivided by the type of assays as, sequencing, CNA, Mutation, Methylation.
}
\examples{
 # Create CGDS object
 cgds<-CGDS("http://www.cbioportal.org/")
 # Get list of cancer studies at server
 Studies <- getCancerStudies(cgds)[,2]
 # Get available case lists (collection of samples) for a given cancer study
 mycancerstudy <- getCancerStudies(cgds)[2,1]
 mycaselist <- getCaseLists(cgds,mycancerstudy)[1,1]
\dontrun{
  ##getCases()
  }
  
}
