% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCasesGenProfs.R
\name{getCasesGenProfs}
\alias{getCasesGenProfs}
\title{get Cases and Genetic Profiles of selected Studies.}
\usage{
getCasesGenProfs()
}
\value{
This function is run by the "Get Cases and Genetic Profiles for selected Studies in starting window. This function needs to select at least one study and display Cases and genetic profiles in the main window.
}
\description{
get Cases and Genetic Profiles of selected Studies.
}
\examples{
cgds <- cBioPortal(
hostname = "www.cbioportal.org",
protocol = "https",
api = "/api/v2/api-docs"
)
\dontrun{
getDataByGenes( api =  cgds,
studyId = "gbm_tcga_pub",
genes = c("NF1", "TP53", "ABL1"),
by = "hugoGeneSymbol",
molecularProfileIds = "gbm_tcga_pub_mrna"
)
}
}
