\name{getCasesGenProfs}
\alias{getCasesGenProfs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get Cases and Genetic Profiles of selected Studies.
}
\description{
This function is run by the "Get Cases and Genetic Profiles for selected Studies in starting window. This function needs to select at least one study and display Cases and genetic profiles in the main window.
}

\examples{
## Not Run
##Load Session
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
## load Cases and Genetic Profiles 
##getCasesGenProfs()
## End(Not run)
}