% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircos.R
\name{getCircos}
\alias{getCircos}
\title{get Circos Layout for selected studies and selected dimensions}
\usage{
getCircos(dimension)
}
\arguments{
\item{dimension}{string (All,mRNA, CNA, Met,RPPA, miRNA, Mut)}
}
\value{
a plot with Circos style
}
\description{
get Circos Layout for selected studies and selected dimensions
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/Circos.rds", sep=""))
\dontrun{
getCircos(dimension ="All")
}
}
