% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicData_MultipleCases.R
\name{getClinicData_MultipleCases}
\alias{getClinicData_MultipleCases}
\title{get Clinical Data for Multiple Cases. User needs to select at least one case to run this function. Get clinical data for more one or multiple cases.}
\usage{
getClinicData_MultipleCases(getSummaryGSEAExists)
}
\arguments{
\item{getSummaryGSEAExists}{if equal to 0, the clinical data is displayed in table. if the argument is equal to 1, the clinical data is used to summarise GSEA analysis results.}
}
\value{
dataframe with clinical data
}
\description{
get Clinical Data for Multiple Cases. User needs to select at least one case to run this function. Get clinical data for more one or multiple cases.
}
\examples{
##Load Session
readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
## Select Case
ENV <- new.env(parent = emptyenv())
ENV$curselectCases <- 2
## get Clinical data
\dontrun{
getClinicData_MultipleCases(getSummaryGSEAExists = 0)
}

}
