% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCor_ExpCNAMet.R
\name{getCor_ExpCNAMet}
\alias{getCor_ExpCNAMet}
\title{Get gene correlation for multiple dimensions.}
\usage{
getCor_ExpCNAMet(ListMatrix, dimension)
}
\arguments{
\item{ListMatrix}{is a List of numeric matrices}

\item{dimension}{Exp,CNA, Met , miRNA , RPPA}
}
\value{
correlation matrix
}
\description{
Get gene correlation for multiple dimensions.
}
\examples{
load(paste(path.package("canceR"),"/data/Circos.RData", sep=""))
\dontrun{
getListProfData()
getCor_ExpCNAMet(myGlobalEnv$ListProfData$Expression, dimension="mRNA")
head(myGlobalEnv$Cor_Exp)
}
}
