\name{geteSet}
\alias{geteSet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Build expression Set (Biobase)
}
\description{
Built Expression Set (eSet) from profile data.
}
\usage{
geteSet()
}

\examples{

\dontrun{
library(canceR)
##load(paste(path.package("canceR"),"/data/brca_tcgaGSEAlm1021.RData", sep=""))
##geteSet()
}
}