% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCT_CLSfiles.R
\name{getGCT_CLSfiles}
\alias{getGCT_CLSfiles}
\title{get Profile (GCT file) and Phenotype (CLS file) Data from Disease.}
\usage{
getGCT_CLSfiles()
}
\value{
GCT and CLS files paths
}
\description{
get Profile (GCT file) and Phenotype (CLS file) Data from Disease.
}
\examples{
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
\dontrun{
getGCT_CLSfiles()
}
}
