\name{getGeneExpMatrix}
\alias{getGeneExpMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
load matrix from file
}
\description{
load matrix from file
}
\usage{
getGeneExpMatrix()
}
%- maybe also 'usage' for other objects documented here.
\value{
Value
}
\examples{

## Not run
library(canceR)
#getGeneExpMatrix()
## End(Not run)
}