\name{getGeneList}
\alias{getGeneList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get Gene List from file.
}
\description{
User needs to specify which gene is interesting to get genomic cancer data. The gene must be with Symbol and one gene by line.
}

\examples{
## Not run
library(canceR)
##getGeneList()
## End(Not run)
}