% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneListFromMSigDB.R
\name{getGeneListFromMSigDB}
\alias{getGeneListFromMSigDB}
\title{get gene list from MSigDB}
\usage{
getGeneListFromMSigDB()
}
\value{
a vector with gene list
}
\description{
get gene list from MSigDB
}
\examples{
load(paste(path.package("canceR"),"/data/brca_tcgaGSEAlm1021.RData", sep=""))
\dontrun{
getGeneListFromMSigDB()
}

}

