% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesTree_MultipleCases.R
\name{getGenesTree_MultipleCases}
\alias{getGenesTree_MultipleCases}
\title{Get  successively trees of genes list for multiple cases}
\usage{
getGenesTree_MultipleCases(entryWidth = 10)
}
\arguments{
\item{entryWidth}{10}
}
\value{
plot tree
}
\description{
Get  successively trees of genes list for multiple cases
}
\examples{
 q <- readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
\dontrun{
readRDS(paste(.libPaths(),"/canceR/data/brca_tcga73genes.rds", sep=""))
getGenesTree_MultipleCases(entryWidth = 10)
}
}
