% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesTree_SingleCase.R
\name{getGenesTree_SingleCase}
\alias{getGenesTree_SingleCase}
\title{classify genes in tree for two phenotypes in the same case(disease).}
\usage{
getGenesTree_SingleCase()
}
\value{
tree plot
}
\description{
classify genes in tree for two phenotypes in the same case(disease).
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/prad_michPhenoTest1021.rds", sep=""))
\dontrun{
getGenesTree_SingleCase()
}
}
