% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenProfs.R
\name{getGenProfs}
\alias{getGenProfs}
\title{Get Genetic Profile from selected Studies}
\usage{
getGenProfs()
}
\value{
dataframe with genetic profil
}
\description{
Get Genetic Profile from selected Studies
}
\examples{
cgds<-CGDS("http://www.cbioportal.org/public-portal/")
# Get list of cancer studies at server
Studies <- getCancerStudies(cgds)[,2]
# Get available case lists (collection of samples) for a given cancer study
mycancerstudy <- getCancerStudies(cgds)[2,1]
mycaselist <- getCaseLists(cgds,mycancerstudy)[1,1]
# Get available genetic profiles
mygeneticprofile <- getGeneticProfiles(cgds,mycancerstudy)[4,1]
\dontrun{
getGenProfs()
}
}

