\name{getGSEAlm_Diseases}
\alias{getGSEAlm_Diseases}

\title{
Get Gene Set Enrichment Analysis linear modeling for two  Diseases.
}
\description{
Get Gene Set Enrichment Analysis linear modeling for two phenotypes (Diseases). Users needs to browse MSigDB file from  ~R/x86_64-unknown-linux-gnu-library/3.2/canceR/extdata/MSigDB/c2.cp.reactome.v4.0.symbols.gmt.
}

\usage{
getGSEAlm_Diseases()
}


\examples{
\dontrun{
##library(canceR)
##library(GSEAlm)
##Load Session
##load(paste(path.package("canceR"),"/data/brca_tcgaGSEAlm1021.RData", sep=""))
##getGSEAlm_Diseases()
}
}