% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGSEAlm_Variables.R
\name{getGSEAlm_Variables}
\alias{getGSEAlm_Variables}
\title{get GSEA linear modeling by variables (phenotype)}
\usage{
getGSEAlm_Variables()
}
\value{
a dataframe with annotation (GO, BP)
}
\description{
get GSEA linear modeling by variables (phenotype)
}
\examples{
 x <- 3
\dontrun{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
getGSEAlm_Variables()
}
}
