% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListProfData.R
\name{getListProfData}
\alias{getListProfData}
\title{get a list of Profile Data of every available dimensions. This function load matrices of every dimension (Exp, CNA, Met, RPPA,miRNA,Mut) and save them in a list for every disease.}
\usage{
getListProfData()
}
\value{
a list of data frame with Profiles Data
}
\description{
get a list of Profile Data of every available dimensions. This function load matrices of every dimension (Exp, CNA, Met, RPPA,miRNA,Mut) and save them in a list for every disease.
}
\examples{
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
\dontrun{
getListProfData() 
head(myGlobalEnv$ProfData$Expression) 
}
}

