\name{getMSigDBExample}
\alias{getMSigDBExample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getMSigDBExample
}
\description{
get example of .gmt file from MSigDB (Broad Institute)
}
\usage{
getMSigDBExample()
}
%- maybe also 'usage' for other objects documented here.
\details{
http://www.broadinstitute.org/gsea/index.jsp
}

\value{
Value
}
\examples{

\dontrun{
library(canceR)
getMSigDBExample()
}
}