% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutData.R
\name{getMutData}
\alias{getMutData}
\title{get Mutation data for multiple genes}
\usage{
getMutData()
}
\value{
a a dataframe with mutation informations
}
\description{
get Mutation data for multiple genes
}
\examples{
load("/Volumes/DATA/canceR/data/ucec_tcga_pubGSEA1021.RData")
\dontrun{
getMutData()
}

}

