\name{getPhenoTest}
\alias{getPhenoTest}

\title{
Associate Phenotypes (variables) from Clinical data to gene expression profile.
}
\description{
see phenoTest package
}


\examples{
\dontrun{
#library(canceR)
#library(Biobase)
#library(phenoTest)
## Load workspace
#load(paste(path.package("canceR"),"/data/prad_michPhenoTest1021.RData", sep=""))

#getPhenoTest()
## Empty Survival raviables
## Select DSF_STATUS for categorical variable
## Select PSA Level for Numeric variale
## BH 
}
}