% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfilesDataSingleGene.R
\name{getProfilesDataSingleGene}
\alias{getProfilesDataSingleGene}
\title{get Profiles Data  for a Single Gene.}
\usage{
getProfilesDataSingleGene()
}
\value{
dataframe with profiles data for a single gene
}
\description{
get Profiles Data  for a Single Gene.
}
\examples{
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
## Select Case from Breast Cancer
myGlobalEnv$curselectCases <- 9
##Select Genetic Profile from Breast Cancer
myGlobalEnv$curselectGenProfs <- 4
## get Specific Mutation data for 73 Genes list
\dontrun{
getProfilesDataSingleGene()
}
}

