% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfilesDataSingleGene.R
\name{getProfilesDataSingleGene}
\alias{getProfilesDataSingleGene}
\title{get Profiles Data  for a Single Gene.}
\usage{
getProfilesDataSingleGene()
}
\value{
dataframe with profiles data for a single gene
}
\description{
get Profiles Data  for a Single Gene.
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/brca_tcga73genes.rds", sep=""))
## Select Case from Breast Cancer
 ENV <- new.env(parent = emptyenv())
ENV$curselectCases <- 9
##Select Genetic Profile from Breast Cancer
ENV$curselectGenProfs <- 4
## get Specific Mutation data for 73 Genes list
\dontrun{
getProfilesDataSingleGene()
}
}
