% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryGSEA.R
\name{getSummaryGSEA}
\alias{getSummaryGSEA}
\title{get Summary  results from GSEA-R (Broad Institute)}
\usage{
getSummaryGSEA()
}
\value{
Dataframe with summary results
}
\description{
get Summary  results from GSEA-R (Broad Institute)
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
Run.GSEA()
getSummaryGSEA()
}
}
