% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvival.R
\name{getSurvival}
\alias{getSurvival}
\title{Survival plot}
\usage{
getSurvival(Coxph)
}
\arguments{
\item{Coxph}{if Coxph = 0 : plot Kaplan-Meier curves else Coxph= 1 : plot Cox Proportional Hazard Model}
}
\value{
Survival plot
}
\description{
Survival plot
}
\examples{
surv <- 11
\dontrun{
load(paste(.libPaths(),"/canceR/data/gbm_tcgaPlotTwoGenProf.RData", sep=""))
getSurvival(Coxph = 1)
}
}
