\name{GSEA.HeatMapPlot}
\alias{GSEA.HeatMapPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
GSEA.HeatMapPlot
}

\description{
GSEA.HeatMapPlot
}
\value{
GSEA.HeatMapPlot
}

\examples{
\dontrun{
library(canceR)
## Load workspace
load(paste(path.package("canceR"),"/data/ucec_tcga_pubGSEA1021.RData", sep=""))
##Run.GSEA()
}
}