% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match_GeneList_MSigDB.R
\name{Match_GeneList_MSigDB}
\alias{Match_GeneList_MSigDB}
\title{Search MSigDb that overlap gene list}
\usage{
Match_GeneList_MSigDB
}
\value{
GeneList
}
\description{
Search MSigDb that overlap gene list
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/prad_michPhenoTest1021.rds", sep=""))
\dontrun{
Match_GeneList_MSigDB()
}
}
