% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalDialog.R
\name{modalDialog}
\alias{modalDialog}
\title{Dialog box to specify Gene Symbol.}
\usage{
modalDialog(title, question, entryInit, entryWidth = 40,returnValOnCancel = "ID_CANCEL")
}
\arguments{
\item{title}{string}

\item{question}{string}

\item{entryInit}{entryInit}

\item{entryWidth}{40}

\item{returnValOnCancel}{"ID_CANCEL"}
}
\value{
dialog box
}
\description{
Dialog box to specify Gene Symbol.
}
\examples{
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
## Select Case from Breast Cancer
myGlobalEnv$curselectCases <- 9
##Select Genetic Profile from Breast Cancer
myGlobalEnv$curselectGenProfs <- 4
## get Specific Mutation data for 73 Genes list
\dontrun{
getProfilesDataSingleGene()
}
}

