\name{plot_1Gene_2GenProfs}
\alias{plot_1Gene_2GenProfs}

\title{
Plotting two genetic profiles for  one Gene 
}
\description{
Tcltk plotting of 1 gene depending two genetic profiles (expression vs mutation or exprssion vs methylation)
}
\usage{
plot_1Gene_2GenProfs()
}

\examples{
\dontrun{
##library(canceR)
## Load workspace
##load(paste(path.package("canceR"),"/data/gbm_tcgaPlotTwoGenProf.RData", sep=""))
##plot_1Gene_2GenProfs()
##use default options
}
}