% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2Genes_1GenProf.R
\name{plot_2Genes_1GenProf}
\alias{plot_2Genes_1GenProf}
\title{plot correlation of two genes expressions.}
\usage{
plot_2Genes_1GenProf()
}
\value{
plot
}
\description{
plot correlation of two genes expressions.
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
plot_2Genes_1GenProf()
}
}
