% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.na.R
\name{rbind.na}
\alias{rbind.na}
\title{bind non equal row}
\usage{
rbind.na(..., deparse.level = 1)
}
\arguments{
\item{...}{...}

\item{deparse.level}{1}
}
\value{
a data frame with merged rows
}
\description{
bind non equal row
}
\examples{

row1 <- c("a","b","c","d")
row2 <- c("A", "B", "C")
row3 <- rbind.na(row1, row2)



}

