% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run.GSEA.1.0.R
\name{Run.GSEA}
\alias{Run.GSEA}
\title{The  main function to run GSEA-R from Broad Institute}
\usage{
Run.GSEA()
}
\value{
A vector with sampling size
}
\description{
The  main function to run GSEA-R from Broad Institute
}
\examples{
readRDS(paste(path.package("canceR"),"/extdata/rdata/ucec_tcga_pubGSEA1021.rds", sep=""))
\dontrun{
Run.GSEA()
}
}
