\name{setWorkspace}
\alias{setWorkspace}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Setting work Directory and output folders.
}
\description{
At starting window, user needs to set work directory for output data. The function is foud in File menu.  
}


\examples{
\dontrun{
library(canceR)
setWorkspace()
}
}


