\name{testCheckedCaseGenProf}
\alias{testCheckedCaseGenProf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Testing checked appropriate Cases for appropriate Genetic profiles.
}
\description{
Testing checked appropriate Cases for appropriate Genetic profiles.
}
\usage{
testCheckedCaseGenProf()
}



\examples{
\dontrun{
##Load Session
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
## Select Case from Breast Cancer
##myGlobalEnv$curselectCases <- 9
##Select Genetic Profile from Breast Cancer
##myGlobalEnv$curselectGenProfs <- 4
##testCheckedCaseGenProf()
}
}