% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCheckedCaseGenProf.R
\name{testCheckedCaseGenProf}
\alias{testCheckedCaseGenProf}
\title{Testing checked appropriate Cases for appropriate Genetic profiles.}
\usage{
testCheckedCaseGenProf()
}
\value{
dialog box with warning message
}
\description{
Testing checked appropriate Cases for appropriate Genetic profiles.
}
\examples{
load(paste(path.package("canceR"),"/data/brca_tcga73genes.RData", sep=""))
\dontrun{
testCheckedCaseGenProf()
}
}

